flask/requirements.txt
Flask
Flask-SQLAlchemy
waitress
python-dotenv